fix(test): configure git user in TestCommitChangelog for CI
The test relied on global git config for author identity, which is
not available in CI environments.
